% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drupal_prep.R
\name{drupal}
\alias{drupal}
\title{Prepare blog post for UVA StatLab Drupal site}
\usage{
drupal(file, name, date, title)
}
\arguments{
\item{file}{an HTML file knitted from R Markdown}

\item{name}{author full name in quotes (e.g., "Clay Ford")}

\item{date}{date in quotes (e.g., "May 30, 2025")}

\item{title}{title of author in quotes (e.g., "Statistical Research Consultant")}
}
\value{
An HTML file of the same name as the file argument with "Drupal_" prepended. Will be written to the current working directory. Open the file with a text editor and copy and paste contents into a new Drupal post.
}
\description{
\code{drupal()} takes an HTML file generated from an R Markdown file in RStudio and makes various changes to the code so it can be copied and pasted into the UVA StatLab Drupal site. This should work for files with both R and Python code chunks.

IMPORTANT: The YAML header of the Rmd file should have output set as follows:

\preformatted{
output:
  html_document:
    self_contained: no
    pandoc_args: ["--wrap=none"]}
}
\examples{
\dontrun{
drupal(file = "tweedie.html",
       name = "Clay Ford",
       date = "May 30, 2025",
       title = "Statistical Research Consultant")
# Drupal_tweedie.html output to working directory.
}
}
