% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statlab_article_metadata.R
\name{get_statlab_article_metadata}
\alias{get_statlab_article_metadata}
\title{Retrieve metadata for all StatLab articles}
\usage{
get_statlab_article_metadata()
}
\value{
A data frame
}
\description{
Collects the article names, authors, and tags for all articles published on the StatLab website at a given time.
}
\examples{
\dontrun{get_statlab_article_metadata()}

}
