% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2.R
\name{hist2}
\alias{hist2}
\title{Histogram with Rug}
\usage{
hist2(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
Histogram with rug added
}
\description{
This creates a histogram with a rug.
}
\details{
The code for rug is \code{rug()}
\itemize{
\item one thing
\item another
}
}
\examples{
hist2(rnorm(100))
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[=hist]{hist()}}
}
