% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{placekick}
\alias{placekick}
\title{NFL placekicking data}
\format{
A data frame with 1425 observations and 9 variables:
\describe{
\item{week}{week of the season}
\item{distance}{distance of the placekick in yards}
\item{change}{does result of kick change who's leading? 0 = no, 1 = yes}
\item{elap30}{number of minutes remaining before end of half; overtime placekicks receive a 0}
\item{PAT}{point after attempt? 0 = no, 1 = yes}
\item{type}{dome = 0, outdoor = 1}
\item{field}{turf = 0, grass = 1}
\item{wind}{no wind = 0, windy = 1; windy is a wind stronger than 15 MPH}
\item{good}{was kick successful? 0 = no, 1 = yes}

}
}
\source{
\url{http://www.chrisbilder.com/categorical/1stEdition/programs_and_data.html}
}
\usage{
placekick
}
\description{
Data on the results of placekicks in the NFL. This is a good data set for demonstrating logistic regression.
}
\examples{
m <- glm(good ~ distance, data = placekick, family = binomial)
exp(confint(m))
}
\references{
Bilder CR and Loughin TM. (2015) \emph{Analysis of Categorical Data with R}. CRC Press.
}
\keyword{datasets}
