% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consultation_report.R
\name{report}
\alias{report}
\title{Generate quick report of StatLab consultation statistics}
\usage{
report(file)
}
\arguments{
\item{file}{a CSV file exported from the LibInsight dataset "5 - RDS+SNE Consultations & Collaborations"}
}
\value{
an HTML report in the current working directory. The Rmd file is in the package inst folder under the rmd subdirectory.
}
\description{
\code{report()} takes a CSV file exported from LibInsight dataset "5 - RDS+SNE Consultations & Collaborations" and creates a quick HTML report of various consultation statistics in the current working directory. The Rmd file requires the lubridate package be installed.
}
\examples{
\dontrun{
report("dataset-17737-20211214162235.csv")
# statlab_consultations_report.html output to working directory.
}
}
