% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{shade}
\alias{shade}
\title{Shade area under standard normal curve}
\usage{
shade(p, lower.tail = TRUE)
}
\arguments{
\item{p}{proportion to shade \verb{(0,1)}}

\item{lower.tail}{logical; if TRUE (default), shading is in the left (or lower)
tail. Otherwise shading is in the right (or upper) tail.}
}
\value{
plot
}
\description{
Specify proportion of area to shade under a standard normal curve. This can
be useful to visualize how probabilities are calculated with a normal
distribution.
}
\details{
The plot title shows the z value corresponding to the area to be shaded. To
find the z-value manually, use the \code{qnorm()} function with p. For example, to
see the z value for p = 0.05, where shading is in the lower tail, run
\code{qnorm(0.05)}. To see the z value for p = 0.05, where shading is in the upper
tail, run \code{qnorm(0.05, lower.tail = FALSE)}.
}
\examples{
shade(p = 0.05)
shade(p = 0.05, lower.tail = FALSE)
}
