% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_statlab1.R
\name{theme_statlab1}
\alias{theme_statlab1}
\title{Create a ggplot2 theme for StatLab}
\usage{
theme_statlab1(
  title_face = "bold",
  title_size = 12,
  axis_color = "black",
  background_color = "white",
  paper = "white",
  ink = "black",
  base_family = "sans",
  base_size = 12,
  text_size = 10,
  base_line_size = 0.5,
  base_rect_size = 0.5,
  use_palette = FALSE
)
}
\arguments{
\item{title_face}{Character. Font for plot title}

\item{axis_color}{Character. Color for axis text and lines}

\item{background_color}{Character. Fill color for the plot background}

\item{paper}{Character. Fill color for panel background}

\item{ink}{Character. Color for all text elements}

\item{base_family}{Character. Font family for all text elements}

\item{base_size}{Numeric. Base font size for text elements}

\item{text_size}{Numeric. Font size for axis text, legend text, and caption}

\item{base_line_size}{Numeric. Base line width for plot lines}

\item{base_rect_size}{Numeric. Border width for rectangular elements}

\item{use_palette}{Logical. TRUE for adding color statlab_palette}
}
\value{
A StatLab ggplot2 theme that can be added to a ggplot2 object
}
\description{
This function applies a theme to ggplot2 plots.
}
\examples{
library(ggplot2)
#create ggplot2
p1 <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
#add statlab_theme1 to ggplot2
p1 + theme_statlab1()

}
\seealso{
\itemize{
\item \href{https://ggplot2.tidyverse.org/}{ggplot2 website} - Official ggplot2 documentation
}
}
