% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wheat}
\alias{wheat}
\title{Wheat kernels}
\format{
A data frame with 175 observations and 7 variables:
\describe{
\item{class}{class of wheat: hard red winter (hrw) or soft red winter (srw)}
\item{density}{density of the kernel (numeric)}
\item{hardness}{hardness of the kernel (numeric)}
\item{size}{size of the kernel (numeric)}
\item{weight}{weight of the kernel (numeric)}
\item{moisture}{moisture of the kernel (numeric)}
\item{type}{condition of kernel: healthy, sprout or scab}

}
}
\source{
\url{http://www.chrisbilder.com/categorical/1stEdition/programs_and_data.html}
}
\usage{
wheat
}
\description{
Data on harvested wheat kernels. This is a good data set for demonstrating multinomial logistic regression.
}
\examples{
head(wheat)
}
\references{
Bilder CR and Loughin TM. (2015) \emph{Analysis of Categorical Data with R}. CRC Press.
}
\keyword{datasets}
