% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statlab_article_metadata.R
\name{get_statlab_article_metadata}
\alias{get_statlab_article_metadata}
\title{Get StatLab Article Metadata}
\usage{
get_statlab_article_metadata()
}
\value{
A data frame of article names, authors, links, and tags.
}
\description{
Collects the article names, authors, links, and tags of all currently published \href{https://library.virginia.edu/data/articles}{StatLab articles}. Function written in 2025 for StatLab website built on drupal.
}
\examples{
\dontrun{d <- get_statlab_article_metadata()}
}
\author{
Ethan Kadiyala
}
